Fix -Wall -Wextra warnings
There is one warning left in "platform.cpp" about GetEntryType(...) where I'm not sure how to solve it.

Removing it breaks the build on platforms that do not provide d_type.

+++ src/game_interpreter.cpp
@@ -3173,8 +3173,8 @@ bool Game_Interpreter::CommandKeyInputProc(lcf::rpg::EventCommand const& com) { 
 		}
 	};
 
-	_keyinput.keys[Keys::eDecision] = check_key(3, true);
-	_keyinput.keys[Keys::eCancel] = check_key(4, true);
+	_keyinput.keys[Keys::eDecision] = check_key(3u, true);
+	_keyinput.keys[Keys::eCancel] = check_key(4u, true);
 
 	// All engines support older versions of the command depending on the
 	// length of the parameter list
@@ -3190,10 +3190,10 @@ bool Game_Interpreter::CommandKeyInputProc(lcf::rpg::EventCommand const& com) { 
 		} else {
 			// For Rpg2k >=1.50
 			_keyinput.keys[Keys::eShift] = com.parameters[5] != 0;
-			_keyinput.keys[Keys::eDown] = check_key(6);
-			_keyinput.keys[Keys::eLeft] = check_key(7);
-			_keyinput.keys[Keys::eRight] = check_key(8);
-			_keyinput.keys[Keys::eUp] = check_key(9);
+			_keyinput.keys[Keys::eDown] = check_key(6u);
+			_keyinput.keys[Keys::eLeft] = check_key(7u);
+			_keyinput.keys[Keys::eRight] = check_key(8u);
+			_keyinput.keys[Keys::eUp] = check_key(9u);
 		}
 	} else {
 		if (param_size != 10 || Player::IsRPG2k3Legacy()) {
@@ -3204,19 +3204,19 @@ bool Game_Interpreter::CommandKeyInputProc(lcf::rpg::EventCommand const& com) { 
 				_keyinput.keys[Keys::eRight] = true;
 				_keyinput.keys[Keys::eUp] = true;
 			}
-			_keyinput.keys[Keys::eNumbers] = check_key(5);
-			_keyinput.keys[Keys::eOperators] = check_key(6);
+			_keyinput.keys[Keys::eNumbers] = check_key(5u);
+			_keyinput.keys[Keys::eOperators] = check_key(6u);
 			_keyinput.time_variable = param_size > 7 ? com.parameters[7] : 0; // Attention: int, not bool
-			_keyinput.timed = check_key(8);
+			_keyinput.timed = check_key(8u);
 			if (param_size > 10 && Player::IsMajorUpdatedVersion()) {
 				// For Rpg2k3 >=1.05
 				// ManiacPatch Middle & Wheel only handled for 2k3 Major Updated,
 				// the only version that has this patch
-				_keyinput.keys[Keys::eShift] = check_key(9, true);
-				_keyinput.keys[Keys::eDown] = check_key(10, true);
-				_keyinput.keys[Keys::eLeft] = check_key(11);
-				_keyinput.keys[Keys::eRight] = check_key(12);
-				_keyinput.keys[Keys::eUp] = check_key(13, true);
+				_keyinput.keys[Keys::eShift] = check_key(9u, true);
+				_keyinput.keys[Keys::eDown] = check_key(10u, true);
+				_keyinput.keys[Keys::eLeft] = check_key(11u);
+				_keyinput.keys[Keys::eRight] = check_key(12u);
+				_keyinput.keys[Keys::eUp] = check_key(13u, true);
 			}
 		} else {
 			// Since RPG2k3 1.05
@@ -3244,11 +3244,11 @@ bool Game_Interpreter::CommandKeyInputProc(lcf::rpg::EventCommand const& com) { 
 #endif
 			return result;
 		};
-		_keyinput.keys[Keys::eMouseLeft] = check_mouse(3);
-		_keyinput.keys[Keys::eMouseRight] = check_mouse(4);
-		_keyinput.keys[Keys::eMouseMiddle] = check_mouse(9);
-		_keyinput.keys[Keys::eMouseScrollDown] = check_mouse(10);
-		_keyinput.keys[Keys::eMouseScrollUp] = check_mouse(13);
+		_keyinput.keys[Keys::eMouseLeft] = check_mouse(3u);
+		_keyinput.keys[Keys::eMouseRight] = check_mouse(4u);
+		_keyinput.keys[Keys::eMouseMiddle] = check_mouse(9u);
+		_keyinput.keys[Keys::eMouseScrollDown] = check_mouse(10u);
+		_keyinput.keys[Keys::eMouseScrollUp] = check_mouse(13u);
 	}
 
 	if (_keyinput.wait) {
@@ -3722,7 +3722,7 @@ bool Game_Interpreter::CommandEndLoop(lcf::rpg::EventCommand const& com) { // co
 		int type = com.parameters[0];
 		int offset = com.indent * 2;
 
-		if (frame.maniac_loop_info.size() < (offset + 1) * 2) {
+		if (static_cast<int>(frame.maniac_loop_info.size()) < (offset + 1) * 2) {
 			frame.maniac_loop_info.resize((offset + 1) * 2);
 			frame.maniac_loop_info_size = frame.maniac_loop_info.size() / 2;
 		}
